changed the styling of his template pages
I added the dawnAndDusk.js file that's needed to adjust how the pages on my site look depending on the time of day. I made excellent use of Notepad++'s extended Replacement tool. It allowed me to change as many as seventy pages of this type in a matter of minutes.