added a cursor preview feature to this template
a standalone style tag was added that creates a cursor override. I got the idea from a website called "TotallyFreeCursors" and believed it would serve a utilitarian purpose in showcasing my themes. The best part is the trivial size of the cursors makes this easy on my site's bandwidth while enhancing the presentation aspect of each hard link DTP page.