fixed rendering issue for sub-gallery list
on the DeskThemePack page, the sub-gallery list had a weird word wrap scenario that would not render well on QHD screens. I added a new rule to the txtFrame class that addresses this. In addition, I have added a new rule specifically for spanned versions of this class to narrow the impact this rule change might have on similar elements. txtFrame is also used as a submenu for other parts of the site.

Also, on the QHD version, the hover color was not a very visible color. This has been fixed. These changes have been replicated on the daytime version of the css file, which does not have a differing entry.