table shrinkage
the use of an H1 tag inside a table is apparently injurious to SEO. These templates correct this anomaly. The only errors that remain will vanish if the template is utilized and placeholders removed.
In addition, it seems you cannot use comments in between section tags. On certain DTP iterations, the cursor of the theme did not show on pages where the description was above the dedicated style tag.
Because different themes use different wallpapers, the use of a dedicated style tag is required.