Significant design changes
These are error free if entered as is in the validator at W3.org.
Part of the reason my pages might not get a lot of views is because the way my pages are coded break a number of HTML5 rules. I cannot be trusted to build a site for a business unless I can get my own right, and with little to no errors.