added hotlink policy
this page often shows up because people try to download content directly from my site. The reason I use Sync.com is because I already use it for storage and it has more bandwidth.
Bandwidth theft reduces the ability of organic viewers of my site and the bandwidth they need to see it all.