font change
I am no longer using Emoji One. The font was difficult to prevent hotlinking on, largely because it's not recognized as a font by my server. It's upsetting, but the one replacing it is much more recent, and much more lean. And because it's a TTF, it won't be easily downloaded. 
If you want the font, it is here on Github:

https://github.com/eosrei/twemoji-color-font/releases/download/v12.0.1/TwitterColorEmoji-SVGinOT-12.0.1.zip