added complimentary transparent background
now in addition to the gradient, a transparent background is used to compliment the gradient texture. Attribution in the filename and the CSS trick URL is also included. Is in active use at 365stops.ORG now. An inverted variant is used for the nite version.

Requires use of RGBA color flags. All existing CSS files include RGBA data in header comments.