changed second title bar
I am preparing to migrate links as I wind down my use of external providers. This will result in the entire `/dtp` directory being suspended in the not too distant future as I work to reformat links and redirects. Should I fail to do this, the number of 404s on my site could shoot through the roof.